add check some some logic

+++ DAL/Bonus_card.cs
@@ -17,7 +17,7 @@ namespace DAL
         [Key]
         public int number_of_card { get; set; }
 
-        public int? kolvo_bonusov { get; set; }
+        public decimal? kolvo_bonusov { get; set; }
 
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Check> Checks { get; set; }
